bman

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int res, count = 0;
+	unsigned long int res, count = 0;
 
 	res = n ^ m;
